rename [c-a] to [basic-a]
