+++ src/components/base/Alert/b-alert-link.vue
@@ -9,7 +9,7 @@
 <script>
 import util from "@/components/util/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 export default {
   name: "BAlertLink",
   components: { BaseA },
