+++ src/components/base/Breadcrumb/b-breadcrumb.vue
@@ -20,7 +20,7 @@
 <script>
 import util from "@/components/util/index.js"
 
-import BaseA from "@/components/basic/A/c-a.vue"
+import BaseA from "@/components/basic/A/basic-a.vue"
 export default {
   name: "BBreadcrumb",
   components: { BaseA },
