+++ src/components/base/Carousel/b-carousel.vue
@@ -64,7 +64,7 @@ import util from "@/components/util/index.js";
 import BCarouselItem from "./b-carousel-item";
 import srMsg from "@/components/basic/basic-sr-msg.vue";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BCarousel",
