+++ src/components/base/Dropdown/b-dropdown-item.vue
@@ -34,7 +34,7 @@
 <script>
 import util from "@/components/util/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 import BInfo from "@/components/basic/basic-info.vue";
 
 export default {
