+++ src/components/base/List/b-list-item.vue
@@ -24,7 +24,7 @@
 <script>
 import util from "@/components/util/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 import srMsg from "@/components/basic/basic-sr-msg.vue";
 
 export default {
