+++ src/components/base/Navigation/Nav/b-nav-item.vue
@@ -16,7 +16,7 @@
 <script>
 import util from "@/components/util/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BNavItem",
