+++ src/components/basic/Button/basic-button.vue
@@ -46,7 +46,7 @@
 <script>
 import util from "@/components/util/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 import BLoading from '@/components/base/Loading/b-loading.vue'
 import srMsg from "@/components/basic/basic-sr-msg.vue";
