+++ src/pages/Demo/Base/b-toast-demo.vue
@@ -188,7 +188,7 @@
 <script>
 import BToast from "@/components/base/toast/b-toast.vue";
 import BToasts from "@/components/base/toast/b-toasts.vue";
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BToastDemo",
