+++ src/pages/Demo/Base/base-icon.vue
@@ -28,7 +28,7 @@
 <script>
 import config from "@/config/index.js";
 
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BaseIcon",
