+++ src/pages/Demo/Base/basic-a-demo.vue
@@ -125,7 +125,7 @@
 </template>
 
 <script>
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BaseADemo",
