+++ src/pages/Demo/Box/b-swiper-demo.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import BaseA from "@/components/basic/A/c-a.vue";
+import BaseA from "@/components/basic/A/basic-a.vue";
 
 export default {
   name: "BSwiperDemo",
