+++ src/pages/Demo/Content/c-a-demo.vue
@@ -122,7 +122,7 @@
 </template>
 
 <script>
-import BaseA from '@/components/basic/A/c-a.vue'
+import BaseA from '@/components/basic/A/basic-a.vue'
 
 export default {
     name: 'CADemo',
