+++ src/pages/Demo/Content/c-zTree-demo.vue
@@ -19,7 +19,7 @@
 
 <script>
 import data from './data.json'
-import CA from '@/components/basic/A/c-a.vue'
+import CA from '@/components/basic/A/basic-a.vue'
 import tree from "@/components/content/Tree/zTree/c-zTree.vue"
 
 export default {
