+++ src/pages/Demo/Form/b-file-demo.vue
@@ -34,7 +34,7 @@
 
 <script>
 import BFile from '@/components/form/b-file.vue'
-import BaseA from '@/components/basic/A/c-a.vue'
+import BaseA from '@/components/basic/A/basic-a.vue'
 
 export default {
   name: 'BFileDemo',
